decoder: Let decoder modules set dpb size.
As many decoder modules do manage their dpb completely internal and do not
benefit from a picture buffer large enough to hold the dpb in VLC itself, let
the decoders decide about the required dpb size.
As of now the only decoder utilizing the dpb through the VLC picture_buffer is
libmpeg2, which is adapted to announce it's actual dpb size.

Signed-off-by: Julian Scheel <julian@jusst.de>