wip: vout_wrapper: Increase private picture count from 4 to 7
As the omxil deinterlace filter can consume up to 5 pictures in parallel (2 in
the deinterlace filter, 2 headed to the output, 1 being prepared) the desired
count of filter reserved pictures of 3 is not sufficient anymore. Increase it
to 6, and hence the total count to 7.
Maybe this should be redesigned to be configurable depending on the used
filter plugins.